requirements.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
asn1crypto==1.5.1
Brotli==1.0.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.1
cssselect2==0.6.0
Django==4.0.4
django-cors-headers==3.11.0
djangorestframework==3.13.1
fonttools==4.33.3
future==0.18.2
html5lib==1.1
idna==3.3
lxml==4.8.0
mysqlclient==2.1.0
oscrypto==1.3.0
Pillow==9.1.0
pycparser==2.21
pydyf==0.1.2
pyHanko==0.13.0
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
pyphen==0.12.0
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
weasyprint==54.3
webencodings==0.5.1
xhtml2pdf==0.2.7
zopfli==0.2.1
